scraper/requirements.txt
tavily-python==0.1.3
playwright==1.40.0
pandas==2.1.3
python-dotenv==1.0.0
uvicorn==0.24.0
fastapi==0.115.5
anyio==3.7.7